+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next version
 
+- Introduced views bundler: An API that allows you to expose view code to frontend JS for client-side templating.
 - Added a button to the validation error page to display the page anyway and another button to disable the validator entirely until the server restarts.
 - Fixed bug with frontend reload causing it to inject the script tag in the wrong location in some situations.
 - Various dependencies bumped.